feat: refactor desktop and mobile navigation components; add home page
