+++ src/pages/HomePage.tsx
@@ -1,0 +1,3 @@
+export function HomePage() {
+  return <div className="p-6">Home</div>;
+}
